requirements.txt
streamlit
langchain
langchain-community
python-dotenv
openai